*** empty log message ***

+++ node/www/cron.php
@@ -1,6 +1,6 @@
 <?php
 /*  -*- tab-width: 3; indent-tabs-mode: 1; -*-
- * $Id: cron.php,v 1.27 2004/04/29 12:46:01 micsik Exp $
+ * $Id: cron.php,v 1.28 2004/05/24 14:45:37 micsik Exp $
  *
  * Created for the StreamOnTheFly project (IST-2001-32226)
  * Authors: András Micsik, Máté Pataki, Tamás Déri 
@@ -119,7 +119,7 @@ while($entry = $dir->read()) {
   if(is_dir($file))
     continue;
   //if (preg_match('/\.png$/', $entry) || preg_match('/\.m3u$/', $entry)) {
-  if(filemtime($file) < $clearTime) {
+  if(is_file($file) && filemtime($file) < $clearTime) {
     if(!unlink($file))
       logError("could not delete: $file");
   }
